Export of internal ZetaSQL changes.
Significant features:
  Very basic reference implementation support //zetasql/public:evaluator (see README)
  Bazel now supports all version >= (version >=0.25.3)
  Change java directory structure (-> ./java/com/google/zetasql)
  Some improvements to java support.

Changelog:
  - ZetaSQL: Update icu ENV to support static linking. by Andrew Pilloud <apilloud@google.com>
  - ZetaSQL: Make jni native code source configurable. by Andrew Pilloud <apilloud@google.com>
  - Many more java unit tests by Matthew Brown <matthewbr@google.com>
  - Add support for evaluation and formatting to java by Matthew Brown <matthewbr@google.com>
  - ScriptExecutor support for "@@error.message" system varia... by ZetaSQL Team <no-reply@google.com>
  - Mark SimpleColumn constructors as public by Matthew Brown <matthewbr@google.com>
  - Support system variables in the reference implementation'... by ZetaSQL Team <no-reply@google.com>
  - Incorporate system variables in name suggestions. by ZetaSQL Team <no-reply@google.com>
  - Support system variables in the reference implementation'... by ZetaSQL Team <no-reply@google.com>
  - Fixing a bug in resolving system variables that prevents ... by ZetaSQL Team <no-reply@google.com>
  - Add resolver support for the presence of the TABLE clause... by ZetaSQL Team <no-reply@google.com>
  - ZetaSQL: Add missing files by Matthew Brown <matthewbr@google.com>
  - Support system variables in the resolver by ZetaSQL Team <no-reply@google.com>
  - Move system variable definitions used by the analyzer tes... by ZetaSQL Team <no-reply@google.com>
  - ZetaSQL: Cleanup to support llvm by Andrew Pilloud <apilloud@google.com>
  - Update several dependency to allow compilation without fl... by Matthew Brown <matthewbr@google.com>
  - Add validation when system variables are inserted into th... by ZetaSQL Team <no-reply@google.com>
  - Extract logic to generate "Unrecognized Name" error messa... by ZetaSQL Team <no-reply@google.com>
  - Add backticks when suggesting symbols with non-identifier... by ZetaSQL Team <no-reply@google.com>
  - Skeleton resolver implementation for system variables. by ZetaSQL Team <no-reply@google.com>
  - Mark GeographyRef class as final as it can not be inheret... by ZetaSQL Team <no-reply@google.com>
  - Add system variables to AnalyzerOptions. by ZetaSQL Team <no-reply@google.com>
  - This fixes a bug in SQLBuilder I encountered while workin... by ZetaSQL Team <no-reply@google.com>
  - ZetaSQL Change up directory structure to be more 'normal'... by Matthew Brown <matthewbr@google.com>
  - ZetaSQL fix build break (caused by previous fix to previo... by Matthew Brown <matthewbr@google.com>
  - No longer mark FEATURE_V_1_3_NULLS_FIRST_LAST_IN_ORDER_BY... by ZetaSQL Team <no-reply@google.com>
  - ZetaSQL: fix break by Matthew Brown <matthewbr@google.com>
  - Fix typo in APPROX_TOP_SUM. by ZetaSQL Team <no-reply@google.com>
  - Cleanup warnings spewed by gcc8+ by Matthew Brown <matthewbr@google.com>
  - Clarify and update support for commandline flags in ZetaS... by Matthew Brown <matthewbr@google.com>
  - Parse SET statements with system variables. by ZetaSQL Team <no-reply@google.com>
  - Simple CLI to execute queries using the reference impleme... by Matthew Brown <matthewbr@google.com>
  - Add DDL to the one-pager document. by ZetaSQL Team <no-reply@google.com>
  - ZetaSQL build break - spelled isnan->std::isnan by Matthew Brown <matthewbr@google.com>
  - Add basic functionality to reference implementation (incl... by Matthew Brown <matthewbr@google.com>
  - ZetaSQL: Outline of reference_impl tests by Matthew Brown <matthewbr@google.com>
  - ZetaSQL: public:evaluator by Matthew Brown <matthewbr@google.com>
  - Modify file-based scripting unit tests to use a restricte... by ZetaSQL Team <no-reply@google.com>
  - ZetaSQL Update SourceLocation, and support StatusBuilder ... by Matthew Brown <matthewbr@google.com>
  - Fix ZetaSQL build break (referencing non-open source libr... by Matthew Brown <matthewbr@google.com>
  - ZetaSQL: reference_impl:evaluation (very limited support). by Matthew Brown <matthewbr@google.com>
  - ZetaSQL: public/functionsl:common_proto by Matthew Brown <matthewbr@google.com>
  - Update absl version to work around gcc bug by Matthew Brown <matthewbr@google.com>
  - Add support hashing invalid Values. by ZetaSQL Team <no-reply@google.com>
  - ZetaSQL: Slight update StatusBuilder by Matthew Brown <matthewbr@google.com>
  - ZetaSQL: Update statusor with advanced construction by Matthew Brown <matthewbr@google.com>
  - Fix garbage comments in icu.BUILD by Matthew Brown <matthewbr@google.com>
  - Parser support for NULLS FIRST and NULLS LAST with ORDER ... by ZetaSQL Team <no-reply@google.com>
  - Remove outdated comment. by ZetaSQL Team <no-reply@google.com>
  - Passing field name to StructField by value to avoid an ex... by ZetaSQL Team <no-reply@google.com>
  - Restructure parse trees so that ASTScript contains an AST... by ZetaSQL Team <no-reply@google.com>
  - Print all NaN values as "nan". The behavior is the same a... by ZetaSQL Team <no-reply@google.com>
  - Many Internal changes by ZetaSQL Team <no-reply@google.com>

GitOrigin-RevId: 32ea82d6cb10feec9915a08f8f04f57738e84ac0

+++ zetasql/base/general_trie.h
@@ -45,6 +45,7 @@
 #include <vector>
 
 #include "zetasql/base/logging.h"
+#include <cstdint>
 
 namespace zetasql_base {
 
@@ -539,7 +540,7 @@ const T& GeneralTrieImpl<T, NullValuePolicy>::GetDataForMaximalPrefixWithLen(
     // See whether we have a match here
     if ((node->data_ != null_value_instance_) &&
         ((s[next_pos] == '\0') || (is_terminator == nullptr) ||
-         (is_terminator[s[next_pos]]))) {
+         (is_terminator[static_cast<int8_t>(s[next_pos])]))) {
       *chars_matched = next_pos;
       matched_data = &(node->data_);
     }
